Refactor into package modules
