+++ eva/contacts.py
@@ -1,0 +1,111 @@
+import json
+import os
+import re
+import traceback
+
+from .config import CONTACTS_FILE
+
+CONTACT_BOOK = {}
+
+
+def load_contacts():
+    global CONTACT_BOOK
+    try:
+        if not os.path.exists(CONTACTS_FILE):
+            with open(CONTACTS_FILE, 'w', encoding='utf-8') as f:
+                json.dump({}, f)
+            CONTACT_BOOK = {}
+            return
+        with open(CONTACTS_FILE, 'r', encoding='utf-8') as f:
+            file_content = f.read()
+            if not file_content.strip():
+                CONTACT_BOOK = {}
+            else:
+                f.seek(0)
+                loaded_data = json.load(f)
+                if isinstance(loaded_data, dict):
+                    CONTACT_BOOK = loaded_data
+                else:
+                    print(f"AVERTISSEMENT [load_contacts]: Contenu de {CONTACTS_FILE} n'est pas un dictionnaire JSON. Reçu: {type(loaded_data)}. Initialisation vide.")
+                    CONTACT_BOOK = {}
+    except json.JSONDecodeError as e:
+        print(f"ERREUR [load_contacts]: Décodage JSON échoué pour {CONTACTS_FILE}: {e}. CONTACT_BOOK réinitialisé.")
+        CONTACT_BOOK = {}
+    except Exception as e:
+        print(f"ERREUR [load_contacts]: Erreur inattendue: {e}. CONTACT_BOOK réinitialisé.")
+        traceback.print_exc()
+        CONTACT_BOOK = {}
+
+
+def save_contacts():
+    global CONTACT_BOOK
+    try:
+        with open(CONTACTS_FILE, 'w', encoding='utf-8') as f:
+            json.dump(CONTACT_BOOK, f, indent=4, ensure_ascii=False)
+        if os.path.exists(CONTACTS_FILE) and os.path.getsize(CONTACTS_FILE) < 2 and len(CONTACT_BOOK) > 0:
+            print(f"AVERTISSEMENT [save_contacts]: Fichier {CONTACTS_FILE} semble vide après sauvegarde alors que CONTACT_BOOK n'est pas vide!")
+        return True
+    except Exception as e:
+        print(f"ERREUR critique [save_contacts] lors de la sauvegarde dans {CONTACTS_FILE}: {e}")
+        traceback.print_exc()
+        return False
+
+
+def add_contact_to_book(name: str, email: str):
+    global CONTACT_BOOK
+    normalized_name = name.lower().strip()
+    email_addr = email.strip()
+    if not normalized_name:
+        return "Le nom du contact ne peut pas être vide."
+    if not re.match(r"[^@]+@[^@]+\.[^@]+", email_addr):
+        return f"L'adresse e-mail '{email_addr}' ne semble pas valide."
+    CONTACT_BOOK[normalized_name] = {"display_name": name, "email": email_addr}
+    if save_contacts():
+        load_contacts()
+        if normalized_name in CONTACT_BOOK and CONTACT_BOOK[normalized_name]["email"] == email_addr:
+            return f"Contact '{name}' ajouté avec l'email '{email_addr}'."
+        else:
+            print(
+                f"AVERTISSEMENT [add_contact_to_book]: Échec de confirmation du contact '{normalized_name}' après rechargement. CONTACT_BOOK: {CONTACT_BOOK}"
+            )
+            return f"Erreur de vérification après sauvegarde du contact '{name}'."
+    else:
+        load_contacts()
+        return f"Erreur lors de la sauvegarde du contact '{name}'."
+
+
+def get_contact_email(name: str):
+    global CONTACT_BOOK
+    normalized_name = name.lower().strip()
+    contact_info = CONTACT_BOOK.get(normalized_name)
+    if contact_info:
+        return contact_info["email"], contact_info.get("display_name", name)
+    return None, name
+
+
+def list_contacts_from_book():
+    global CONTACT_BOOK
+    if not CONTACT_BOOK:
+        return "Votre carnet d'adresses est vide."
+    return "Voici vos contacts :\n" + "\n".join([f"- {c['display_name']} ({c['email']})" for c in CONTACT_BOOK.values()])
+
+
+def remove_contact_from_book(name: str):
+    global CONTACT_BOOK
+    normalized_name = name.lower().strip()
+    if normalized_name in CONTACT_BOOK:
+        removed_contact_name = CONTACT_BOOK[normalized_name]["display_name"]
+        del CONTACT_BOOK[normalized_name]
+        if save_contacts():
+            load_contacts()
+            if normalized_name not in CONTACT_BOOK:
+                return f"Contact '{removed_contact_name}' supprimé."
+            else:
+                print(f"AVERTISSEMENT [remove_contact_from_book]: Contact '{removed_contact_name}' toujours présent après rechargement.")
+                return f"Erreur de confirmation de suppression pour '{removed_contact_name}'."
+        else:
+            load_contacts()
+            return f"Erreur de sauvegarde après suppression du contact '{removed_contact_name}'."
+    return f"Contact '{name}' non trouvé."
+
+load_contacts()
